refactor: move to auth module
